Allow get('/action', { :input => 'value' }) and get('/action', 'input=value')
darcs-hash:20071107182051-569a4-08296cd576c07492da6eafcc397a91dbb5b02ebf.gz

+++ lib/ramaze/spec/helper/mock_http.rb
@@ -40,7 +40,7 @@ module MockHTTP
   def process_request(path, query)
     options = {}
     FISHING.each{|key, value|
-      options[value] = query.delete(key)}
+      options[value] = query.delete(key)} if query.is_a?(Hash)
     [create_url(path, query), options]
   end
 
@@ -52,9 +52,9 @@ module MockHTTP
   end
 
 	def make_query query
-		return query unless query and query.respond_to?(:map)
+		return query unless query and not query.is_a?(String)
     query.map{|key, value|
-      "#{CGI::escape(key)}=#{CGI::escape(value)}"
+      "#{CGI::escape(key.to_s)}=#{CGI::escape(value.to_s)}"
     }.join('&')
 	end
 end
